requirements.txt
# StrategyBuilder Pro - Dependencies

# Core trading and backtesting
backtrader==1.9.78.123
yfinance==0.2.66

# Data analysis and manipulation
pandas==2.3.3
numpy==2.3.4

# Web framework and UI
streamlit==1.51.0

# Visualization
plotly==6.4.0

# Data fetching dependencies
beautifulsoup4==4.14.2
lxml==6.0.2
html5lib==1.1
curl_cffi==0.13.0
websockets==15.0.1

# yfinance dependencies
frozendict==2.4.7
peewee==3.18.3
platformdirs==4.5.0

# HTTP requests
requests>=2.31.0

# Python timezone handling
pytz>=2022.5

# Date utilities
python-dateutil>=2.8.2

# Configuration
PyYAML>=6.0

# Testing and coverage
coverage>=7.0.0

# Note: multitasking may require manual installation due to setuptools compatibility issues
# If pip install fails for multitasking, you may need to install it manually
